backend/requirements-txt.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-Cors==3.0.10
PyJWT==2.3.0
stripe==2.65.0
